refactor(plugin): delete the dev-server global left over from the runner proxies
Replacing the runner proxies with plugin-rsc's loadModule left
vikeReactRscGlobalState with one field, written once by a plugin whose only
job was to write it, and read nowhere. plugin-rsc acquires environment
runners itself, so the plugin, the global, its type and its initializer all
go. The dev lane still serves every route and the RSC endpoint without them.

The bridge's post-merge check could not fail: the projected object is built
by spreading the root input first, so no root key can be absent from it. It
read as a guard while guarding nothing. The derived exact-set comparison
that does have teeth already runs against the built artifacts.

normalizeRollupInput moves in beside its only caller and stops being
exported. sharedDuringBuild is dropped: Vite consults it only in the
per-environment re-resolution branch, which sharedConfigBuild skips.
