+++ packages/vike-react-rsc/src/plugin/index.ts
@@ -1,29 +1,15 @@
-import { type PluginOption, type ViteDevServer } from "vite";
+import { type PluginOption } from "vite";
 import { clientInputBridge } from "./plugins/clientInputBridge";
 import { configs } from "./plugins/config";
-import { exposeDevServer } from "./plugins/dev";
 import { vikeRscManifestPluginBuild } from "./plugins/injectManifestBuild";
 import { virtuals } from "./plugins/virtuals";
 import rsc from "@vitejs/plugin-rsc";
-
-type GlobalState = {
-  devServer?: ViteDevServer;
-};
-
-declare global {
-  var vikeReactRscGlobalState: GlobalState;
-}
-
-global.vikeReactRscGlobalState ||= {
-  devServer: undefined,
-};
 
 export default function vikeRscPlugin(): PluginOption[] {
   //@ts-ignore
   return [
     ...configs,
     ...virtuals,
-    exposeDevServer,
     vikeRscManifestPluginBuild(),
     clientInputBridge(),
     ...rsc({
