+++ packages/vike-react-rsc/src/plugin/plugins/config.ts
@@ -103,7 +103,6 @@ export const configs: Plugin[] = [
         },
       };
     },
-    sharedDuringBuild: true,
   },
   {
     name: "vike-rsc:config-rsc",
