--- packages/vike-react-rsc/src/plugin/plugins/dev.ts
@@ -1,16 +1,0 @@
-import type { Plugin, RunnableDevEnvironment } from "vite";
-
-export const exposeDevServer: Plugin = {
-  name: "rsc-dev",
-  configureServer(server) {
-    global.vikeReactRscGlobalState.devServer = server;
-    try {
-      if (server.environments.rsc)
-        (server.environments.rsc as RunnableDevEnvironment).runner;
-      if (server.environments.ssr)
-        (server.environments.ssr as RunnableDevEnvironment).runner;
-    } catch (e) {
-      console.error("[RSC Plugin] Failed to initialize runners:", e);
-    }
-  },
-};
