+++ packages/vike-react-rsc/src/plugin/utils.ts
@@ -1,20 +1,6 @@
-import type { Plugin, BuildEnvironmentOptions } from "vite";
+import type { Plugin } from "vite";
 
-export { createVirtualPlugin, normalizeRollupInput };
-
-type RollupInput = NonNullable<
-  NonNullable<BuildEnvironmentOptions["rollupOptions"]>["input"]
->;
-
-function normalizeRollupInput(
-  input: RollupInput | undefined
-): Record<string, string> {
-  if (!input) return {};
-  if (typeof input === "string") return { [input]: input };
-  if (Array.isArray(input))
-    return Object.fromEntries(input.map((entry) => [entry, entry]));
-  return { ...input };
-}
+export { createVirtualPlugin };
 
 // Helper to create virtual plugins
 function createVirtualPlugin(name: string, load: Plugin["load"]): Plugin {
